refactor(router): move Navigator interfaces to navigator
