+++ packages/router/src/browser-navigator.ts
@@ -1,31 +1,8 @@
 import { Key, Reporter } from '@aurelia/kernel';
 import { IDOM, ILifecycle } from '@aurelia/runtime';
 import { HTMLDOM } from '@aurelia/runtime-html';
-import { INavigatorState, IStoredNavigatorEntry } from './navigator';
+import { INavigatorState, INavigatorStore, INavigatorViewer, INavigatorViewerEvent } from './navigator';
 import { Queue, QueueItem } from './queue';
-
-export interface INavigatorStore {
-  length: number;
-  state: Record<string, unknown>;
-  go(delta?: number, suppressPopstate?: boolean): Promise<void>;
-  pushNavigatorState(state: INavigatorState): Promise<void>;
-  replaceNavigatorState(state: INavigatorState): Promise<void>;
-  popNavigatorState(): Promise<void>;
-}
-
-export interface INavigatorViewer {
-  activate(callback: (ev?: INavigatorViewerEvent) => void): void;
-  deactivate(): void;
-}
-
-export interface INavigatorViewerEvent {
-  event: PopStateEvent;
-  state?: INavigatorState;
-  path: string;
-  data: string;
-  hash: string;
-  instruction: string;
-}
 
 interface Call {
   target: object;
